ui: add network logs in assist
